+++ extensions/youtube/src/main/java/app/revanced/extension/youtube/patches/components/LayoutComponentsFilter.java
@@ -247,8 +247,13 @@ public final class LayoutComponentsFilter extends Filter {
                 "sponsorships"
         );
 
+        final var crowdfundingBox = new StringFilterGroup(
+                Settings.HIDE_CROWDFUNDING_BOX,
+                "donation_shelf"
+        );
+
         final var channelWatermark = new StringFilterGroup(
-                Settings.HIDE_VIDEO_CHANNEL_WATERMARK,
+                Settings.HIDE_CHANNEL_WATERMARK,
                 "featured_channel_watermark_overlay"
         );
 
@@ -312,6 +317,7 @@ public final class LayoutComponentsFilter extends Filter {
                 compactChannelBar,
                 compactChannelBarInner,
                 communityPosts,
+                crowdfundingBox,
                 emergencyBox,
                 expandableMetadata,
                 forYouShelf,
@@ -427,7 +433,7 @@ public final class LayoutComponentsFilter extends Filter {
      * Injection point.
      */
     public static boolean showWatermark() {
-        return !Settings.HIDE_VIDEO_CHANNEL_WATERMARK.get();
+        return !Settings.HIDE_CHANNEL_WATERMARK.get();
     }
 
     /**